script_checker/requirements.txt
anthropic>=0.30.0
streamlit>=1.35.0
python-dotenv>=1.0.0
